# opt: use OrderedDict to make LRUCache to optimize efficiency of set o…