@NoneGG 2018-09-08
@msiemens 
There seems to be some failed test cases not related with this change ( 0-0 ）
Cases in `test_utils` all pass tests