@msiemens 2018-09-13
@NoneGG I've rerun the test suite and it seems like the change actually changes the behavior of the LRU cache in a way that makes it return invalid/unexpected values. The test suite also doesn't pass locally for me.